refactor(Console): 입출력 객체 분리
